Change stack to use size_t instead of unsigned int

+++ include/stack.h
@@ -1,14 +1,17 @@
 #ifndef STACK_H
 #define STACK_H
 
+#include <stdint.h>
+#include <stddef.h>
+
 typedef struct{
-    unsigned int size;
-    unsigned int capacity;
+    size_t size;
+    size_t capacity;
     int* data;
 }stack;
 
 int init_stack(stack *s);
-int init_stack_with_capacity(stack *s, int capacity);
+int init_stack_with_capacity(stack *s, size_t capacity);
 void free_stack(stack *s);
 int push(stack *s, int value);
 int pop(stack *s, int *out);
